(wip) move tee proof to web-prover-core
