+++ client/Cargo.toml
@@ -27,8 +27,7 @@ tls-client        ={ workspace=true }
 tls-client2       ={ workspace=true }
 tls-core          ={ workspace=true }
 client-side-prover={ git="https://github.com/pluto/client-side-prover", rev="8e7eb839e901dcee416179116bb0f9c4f7ae683c" }
-
-web-prover-core={ workspace=true }
+web-prover-core   ={ workspace=true }
 
 # TLSN
 tlsn-prover={ workspace=true }
