+++ web-prover-core/src/lib.rs
@@ -2,4 +2,5 @@ pub mod errors;
 pub mod http;
 pub mod manifest;
 
+pub mod proof;
 pub mod test_utils;
